refactor(application): move blocking database helper into collections
